+++ src/main/java/org/sonar/plugins/delphi/core/helpers/DelphiProjectHelper.java
@@ -315,6 +315,10 @@ public class DelphiProjectHelper {
     return fs != null ? fs.encoding().name() : Charset.defaultCharset().name();
   }
 
+  public String testTypeRegex() {
+    return settings.get(DelphiPlugin.TEST_TYPE_REGEX_KEY).orElse("(?!)");
+  }
+
   /**
    * Is file in excluded list?
    *